v2.0.0: stability declaration
Second stability declaration. v1.0 declared retrieval stable; v2.0
declares the full surface stable — chat (local + Claude), public
tunnels (quick + named), external MCP servers, collections.
No breaking REST or MCP changes; v1.x vaults open without migration.

Search + Ask gain a CollectionScope chip row matching the Documents
tab strip. Pulled into a shared component so future routes inherit
the control.

DELETE /api/sources now requires {confirm: true} | {dry_run: true}
matching /api/clear's gate. The web UI auto-supplies confirm on
writes (it already gates writes through a dry-run preview), so
existing flows are unchanged. External API consumers must add the
flag on write calls.

Lib + binary clippy now zero warnings: manual_clamp on the sources
limit, redundant `as usize` in collection_remove_documents, and
allow(too_many_arguments) on db.search after the v1.6.1 collection
scope brought the arg count to eight.

147 lib + 25 integration tests, all green. README refreshed for the
v2.0 surface; tunnel risk note now includes the new write guard.

+++ src/rag/mod.rs
@@ -240,6 +240,7 @@ impl Database {
     ///
     /// `query_text` is what FTS5 tokenizes for keyword matching. Pass the same
     /// natural-language string the user typed; do not pre-tokenize.
+    #[allow(clippy::too_many_arguments)]
     pub fn search(
         &self,
         query_embedding: &[f32],
@@ -839,7 +840,7 @@ impl Database {
                 "DELETE FROM document_collections WHERE collection_id = ?1 AND document_id = ?2",
             )?;
             for doc_id in document_ids {
-                removed += stmt.execute(params![id, doc_id])? as usize;
+                removed += stmt.execute(params![id, doc_id])?;
             }
         }
         tx.commit()?;
